requirements.txt
# Bayesian Networks & ML
pgmpy==0.1.23
pymc==5.0.0
pomegranate==0.14.9
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# Geospatial & GIS
geopandas==0.13.2
shapely==2.0.1
rasterio==1.3.7
fiona==1.9.4
gdal==3.7.0

# Visualization & Web
folium==0.14.0
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
leaflet==0.0.3

# Data & APIs
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
geoalchemy2==0.13.1

# Utilities
pyyaml==6.0
python-dotenv==1.0.0
logging-config==0.4.1

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Jupyter
jupyter==1.0.0
jupyterlab==4.0.5
ipython==8.15.0